requirements.txt
# Flask Web Server Dependencies
# For deployment on Render.com, Heroku, or other cloud platforms

Flask>=3.0.0
Werkzeug>=3.0.0
mysql-connector-python>=8.0.0
paho-mqtt>=1.6.0
pycryptodome>=3.19.0
cryptography>=41.0.0
gunicorn>=21.2.0



